% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.residue.R
\name{com.residue}
\alias{com.residue}
\title{com.residue is a function used to calculate the residue of GARCH model so that the futher adaptive procedure can be move forward.}
\usage{
com.residue(alpha, beta, series)
}
\arguments{
\item{alpha}{Includes intercept and all parameters of ARCH terms.}

\item{beta}{Includes all parameters of GARCH terms.}

\item{series}{Time series.}
}
\value{
Return a list contains estimators, residues and conditional sigma square.
}
\description{
com.residue is a function used to calculate the residue of GARCH model so that the futher adaptive procedure can be move forward.
}
\examples{
# need to be changed.
}

