% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH1_1wNorm.R
\name{GARCH1_1}
\alias{GARCH1_1}
\title{GARCH(1,1)}
\usage{
GARCH1_1(n, a, b)
}
\arguments{
\item{n}{series number}

\item{a}{alpha include intercept and alpha1}

\item{b}{beta}
}
\value{
A list includes GARCH(1,1) series, error series and conditional variances series.
}
\description{
Example to generate GARCH(1,1) with normal innovation.
}
\examples{
# Need revisions.
}

