% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHighFreqTrade.R
\name{getHighFreqData}
\alias{getHighFreqData}
\title{getHighFreqData}
\usage{
getHighFreqData(instrument, start, end, provider, savename)
}
\arguments{
\item{instrument}{The quote symbol of the stock. This quote should correspond with Yahoo list. Symbols can be found here: http://www.marketwatch.com/tools/industry/stocklist.asp?bcind_ind=9537&bcind_period=3mo}

\item{start}{The start date of your period with the format "YYYY-MM-DD". Default start date is "1991-01-02".}

\item{end}{The end date of your period with the format "YYYY-MM-DD". The default end date is the day before the today in your system.}

\item{provider}{The source of data, contains options: "google", "yahoo", "onada".}

\item{savename}{The saving name the file including downloaded data. The default folder is the current working directory of the R process.}
}
\value{
The time series data will be returned.
}
\description{
This function is used to download high frequent data of stock. The source is Google.
}
\examples{
# need revisions
}

