% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.intraday.data.R
\name{google.intraday.data}
\alias{google.intraday.data}
\title{google.intraday.data is used to download Google minutely trading date of stocks.}
\usage{
google.intraday.data(symbol, freq, period)
}
\arguments{
\item{symbol}{The stock symbol. This quote should correspond with Yahoo list. Symbols can be found here: http://www.marketwatch.com/tools/industry/stocklist.asp?bcind_ind=9537&bcind_period=3mo}

\item{freq}{The interval of time. The unit of it is second. The minimal interval is 60 seconds. Like "60s". Please input numeric.}

\item{period}{The time period. The unit is day, "d". Please input numeric.}
}
\value{
Return data include data untill now.
}
\description{
This is a function used get Google minutely trading data of stocks.
}
\examples{
# Need examples.
}

