% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogL_Linear_Norm.R
\name{LogL_Linear_Norm}
\alias{LogL_Linear_Norm}
\title{Log-likelihood function of single linear model.}
\usage{
LogL_Linear_Norm(X, y)
}
\arguments{
\item{X}{The independent variable x without intercept.}

\item{y}{The dependent variable y.}
}
\value{
Log-likelihood function
}
\description{
The Log-likelihood function of linear model. Here is the y=a + b*x. Need to be changed.
}

