% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH_t.R
\name{GARCH_t}
\alias{GARCH_t}
\title{T.GARCH}
\usage{
GARCH_t(alpha, beta, n = 100, rnd = rt, df.t, ntrans = 100)
}
\arguments{
\item{alpha}{The vector include the incept and the parameters of autoregression part}

\item{beta}{The vector include the parameters of conditional variancea}

\item{n}{The lenght of time series}

\item{rnd}{The distribution of error(innovation) of GARCH model}

\item{df.t}{The df of student's t distribution. df.t>0 continuous parameter}

\item{ntrans}{burn-in size, i.e. number of initial simulated data to be discarded}
}
\description{
Function to generate GARCH time series with t-students distribution innovations
}
\details{
This function is used to generate GARCH time series with t-students distribution innovations. This kind of innovations will better perform the heavy-tailed characteristics of financial data.
}
\author{
Xiaorui.Zhu
}

